+++ actix-broker/src/broker.rs
@@ -5,10 +5,9 @@ use std::{
     marker::PhantomData,
 };
 
+use actix::prelude::*;
 use ahash::AHasher;
 use log::trace;
-
-use actix::prelude::*;
 
 use crate::msgs::*;
 
@@ -51,15 +50,20 @@ impl Broker<ArbiterBroker> {
 
 /// The system service actor that keeps track of subscriptions and routes messages to them.
 impl<T> Broker<T> {
-    fn get_subs<M: BrokerMsg>(&self) -> Option<impl Iterator<Item=(usize, (&TypeId, &Recipient<M>))>> {
+    fn get_subs<M: BrokerMsg>(
+        &self,
+    ) -> Option<impl Iterator<Item = (usize, (&TypeId, &Recipient<M>))>> {
         let id = TypeId::of::<M>();
         let msg_subs = self.sub_map.get(&id)?;
 
         trace!("Broker: Found subscription list for {:?}.", id);
 
-        let iter = msg_subs.iter().enumerate()
+        let iter = msg_subs
+            .iter()
+            .enumerate()
             .map(|(idx, (actor_id, recipient))| {
-                let typed_recipient = recipient.downcast_ref::<Recipient<M>>()
+                let typed_recipient = recipient
+                    .downcast_ref::<Recipient<M>>()
                     .expect("Message type should always be M");
                 (idx, (actor_id, typed_recipient))
             });
@@ -87,10 +91,9 @@ impl<T> Broker<T> {
         let msg_id = TypeId::of::<M>();
         if let Some(subscribers) = self.sub_map.get_mut(&msg_id) {
             trace!("Broker: Removing {:?} subscribers", indexes);
-            indexes.iter().rev()
-                .for_each(|&idx| {
-                    subscribers.swap_remove(idx);
-                });
+            indexes.iter().rev().for_each(|&idx| {
+                subscribers.swap_remove(idx);
+            });
         }
     }
 
@@ -141,21 +144,18 @@ impl<T: 'static + Unpin, M: BrokerMsg> Handler<IssueAsync<M>> for Broker<T> {
     fn handle(&mut self, msg: IssueAsync<M>, _ctx: &mut Context<Self>) {
         trace!("Broker: Received IssueAsync");
         let subscriber_indexes = if let Some(subscribers) = self.get_subs::<M>() {
-            subscribers.filter(|&(_, (actor_id, _))| !msg.1.eq(actor_id))
-                .map(|(idx, (_, recipient))| {
-                    match recipient.try_send(msg.0.clone()) {
+            subscribers
+                .filter(|&(_, (actor_id, _))| !msg.1.eq(actor_id))
+                .map(
+                    |(idx, (_, recipient))| match recipient.try_send(msg.0.clone()) {
                         Err(SendError::Full(msg)) => {
                             recipient.do_send(msg);
                             None
                         }
-                        Err(_) => {
-                            Some(idx)
-                        }
-                        _ => {
-                            None
-                        }
-                    }
-                })
+                        Err(_) => Some(idx),
+                        _ => None,
+                    },
+                )
                 .filter(|actor_idx| actor_idx.is_some())
                 .map(|idx| idx.expect("Actor index should always be present"))
                 .collect()
@@ -175,9 +175,11 @@ impl<T: 'static + Unpin, M: BrokerMsg> Handler<IssueSync<M>> for Broker<T> {
         trace!("Broker: Received IssueSync");
 
         if let Some(subscribers) = self.get_subs::<M>() {
-            subscribers.filter(|&(_, (actor_id, _))| !msg.1.eq(actor_id))
+            subscribers
+                .filter(|&(_, (actor_id, _))| !msg.1.eq(actor_id))
                 .for_each(|(_, (_, recipient))| {
-                    recipient.send(msg.0.clone())
+                    recipient
+                        .send(msg.0.clone())
                         .into_actor(self)
                         .map(|_, _, _| ())
                         .wait(ctx);